Introduce an app_root variable to superset:create_app.
If not an empty string, a new middleware class is attached
to the wsgi_app callable that intercepts each request and
transforms the SCRIPT_NAME & PATH_INFO
variables in the request environment so that the rest of
the code still sees non-prefixed paths yet the application
itself appears to sit on a common subpath.

A new environment variable SUPERSET_APP_ROOT has been
introduced into the docker config to control the value
of the application root.